# src/main/java/com/example/OpenAPISpec/dto/StudentsDto.java
package com.example.OpenAPISpec.dto;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class StudentsDto {

    private String name;

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public String getClassName() {
        return className;
    }

    public void setClassName(String className) {
        this.className = className;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public long getPhoneNum() {
        return phoneNum;
    }

    public void setPhoneNum(long phoneNum) {
        this.phoneNum = phoneNum;
    }

    private String address;
    private String className;
    private long phoneNum;


}